+++ src/commands/files/index.ts
@@ -1,2 +1,3 @@
 export { getFile } from './get.js';
+export { listFiles, listFilesFormatted, ListFilesOptions } from './list.js';
 export { uploadFile, UploadFileOptions } from './upload.js';